Limiting to 10 iterations to get a complete message
Replacing a while by a for with range 10. Before the maximum holding
time would be the total bytes in a message times the listening timout.
Now it is capped to at most 10 increments to listen the rest of a
truncated message. In my experience so far, when it happens, it requires
a single time to complete the message.